make some stuff compile
