+++ crates/actors/src/epoch_service/epoch_service.rs
@@ -6,10 +6,7 @@ use irys_config::StorageSubmodulesConfig;
 use irys_database::{data_ledger::*, SystemLedger};
 use irys_primitives::CommitmentStatus;
 use irys_storage::{ie, StorageModuleInfo};
-use irys_types::{
-    partition::{PartitionAssignment, PartitionHash},
-    IrysBlockHeader, SimpleRNG, H256,
-};
+use irys_types::{partition::{PartitionAssignment, PartitionHash}, DataLedger, IrysBlockHeader, SimpleRNG, H256};
 use irys_types::{
     partition_chunk_offset_ie, Address, CommitmentTransaction, ConsensusConfig,
     PartitionChunkOffset,
