+++ crates/api-server/src/routes/price.rs
@@ -4,14 +4,10 @@ use actix_web::{
     HttpResponse, Result as ActixResult,
 };
 use irys_actors::ema_service::EmaServiceMessage;
-use irys_database::DataLedger;
-use irys_types::{
-    storage_pricing::{
-        phantoms::{Irys, NetworkFee},
-        Amount,
-    },
-    U256,
-};
+use irys_types::{storage_pricing::{
+    phantoms::{Irys, NetworkFee},
+    Amount,
+}, DataLedger, U256};
 use serde::{Deserialize, Serialize};
 
 use crate::ApiState;
