+++ crates/api-server/src/routes/tx.rs
@@ -6,10 +6,8 @@ use actix_web::{
 };
 use awc::http::StatusCode;
 use irys_actors::mempool_service::{TxIngressError, TxIngressMessage};
-use irys_database::{database, DataLedger};
-use irys_types::{
-    u64_stringify, CommitmentTransaction, IrysTransactionHeader, IrysTransactionResponse, H256,
-};
+use irys_database::{database};
+use irys_types::{u64_stringify, CommitmentTransaction, DataLedger, IrysTransactionHeader, IrysTransactionResponse, H256};
 use reth_db::Database;
 use serde::{Deserialize, Serialize};
 use tracing::info;
