+++ crates/storage/src/chunk_provider.rs
@@ -1,6 +1,5 @@
 use eyre::OptionExt;
-use irys_database::DataLedger;
-use irys_types::{ChunkFormat, Config, DataRoot, LedgerChunkOffset, PackedChunk, TxChunkOffset};
+use irys_types::{ChunkFormat, Config, DataLedger, DataRoot, LedgerChunkOffset, PackedChunk, TxChunkOffset};
 use std::sync::Arc;
 
 use tracing::debug;
